+++ docs/CHANGELOG.md
@@ -1,8 +1,25 @@
 # Enigma Bot changelog
 
 ## [Unreleased]
+#### Added
+- Logging when bot is online and ready.
+- Logging Discord connection states.
+#### Fixed
+- Logging after leaving a server.
+#### Updated
+- [`setup.py`] now has all packages
+#### Changed
+- Python version in [`setup.py`] is now read from [`runtime.txt`]
+- There's author's full name in [`setup.py`].
+- Name in [`setup.py`] now looks better.
+- `info` command embed.
+- `spotify` command help.
 
 ## Released
+### v0.7-hotfix - 09.02.2021
+#### Fixed
+- `spotify` command was disabled in production.
+
 ### [v0.7] - 08.02.2021
 #### Added
 - `spotify` command.
@@ -107,6 +124,8 @@
 - Communication with the database using SQLAlchemy.
 
 
+[`setup.py`]: https://github.com/AnonymousX86/Enigma-Bot/blob/master/setup.py
+[`runtime.txt`]: https://github.com/AnonymousX86/Enigma-Bot/blob/master/runtime.txt
 [Unreleased]: https://github.com/AnonymousX86/Enigma-Bot/compare/deploy...master
 [v0.7]: https://github.com/AnonymousX86/Enigma-Bot/releases/tag/v0.7
 [v0.6]: https://github.com/AnonymousX86/Enigma-Bot/releases/tag/v0.6